requirements.txt
python-telegram-bot
sqlalchemy
python-dotenv
pre-commit
pyrogram
TgCrypto
